requirements.txt
flask>=2.0
flasgger>=0.9
requests>=2.25
flask_sqlalchemy>=3.0
pymysql>=1.1
Werkzeug~=3.1.3
python-dotenv~=1.1.1
cryptography>=42